New and improved Task
+ Task is now an interface
+ Task implements Runnable interface
